backend/python_services/crewai_service/requirements.txt
crewai==0.114.0
langchain==0.3.23
openai==1.72.0
fastapi==0.115.9
uvicorn==0.34.0
pydantic==2.11.3
python-dotenv==1.1.0
